cockroach-go/testserver: testserver using real binaries
testserver package using a real cockroach binary (latest is
automatically downloaded, with fallback to PATH/cockroach).

Provides: Start/Stop/PGURL.
Trivial tests just to show the usage.

Parsing stdout to find the pgurl is a little gross, but we need a
dynamic port to make sure we don't conflict with running binaries.