Wrap errors when rollback to savepoint
If an error occurs attempting to rollback to a savepoint, and that error
isn't a libpq error, then the error is currently overwritten by a string.
This causes it to lose its structure and any context it had. Instead,
wrapping the error allows callers to be able to deconstruct the error.