add support for pgx and gorm
Before this patch, the library only had support for Go's sql package.
This patch adds subpackages for pgx and gorm, providing suitable
ExecuteTx() functions. We had a function called ExecuteInTx which was
intended to be used with transactions initiated by other frameworks.
However the function's abstraction of a Txn did not match any framework,
and the API was generally dangerous because it simply assumed that the
respective txn had not previously been used to run statements.
In contrast, the new ExecuteTx() functions work with the respective
frameworks and controls the creation of the transaction.

This patch also makes the crdb package not import a pgx package any
more.

Fixes #60