testserver: add option to run cockroach server in secure mode
The server would previously run in insecure mode. This change adds the ability
for test environments to turn secure mode on if desired, which is more in line
with what a production environment would be like.

The PGURL with secure connection parameters is returned through the TestServer
level PGURL method.