support downloading the latest stable version of CRDB
resolves #83

This commit sets the default mode for database downloads
to "Download the latest stable version of CRDB".

We get the latest stable version through
https://register.cockroachdb.com/api/updates, and
save it to a local file with `cockroach-majorv-minorv-patch`
naming pattern in the `tmp` directory. The downloading is
OS-specific.

**Problem unsolved**: the tenant tests with proxy always fail
with stable CRDB, but pass with latest non-stable CRDB.