Add tenant-scope arg for client certs
This PR adds tenant scoping to the client certs
created for accessing tenant servers. Tenant
scopes were added as a way to restrict the ability
of client certs to validate with a specific set of
tenants in v22.2. Without tenant scoping, the client
certs created for tenant servers will not be able to
connect to non system tenant servers. This PR fixes
this problem by updating the client cert generation
command to specify the tenant scope argument while
creating client certs for tenant servers.