Support custom host on listen address
Previously, the testserver could only listen on localhost. This meant
that there was no way to spin up an instance of the testserver on one
machine and access it from another. This commit adds support for custom
hosts so that the testserver can be accessed from outside hosts if
needed.