testserver: respect the TMPDIR environment variable
This will be useful for cockroachdb CI, which has logic that can
preserve temporary files in TMPDIR in the CI artifacts.

It's safe to remove the logic to use "/tmp" since that was initially
added in a2ed1c60, but it's no longer needed since the testserver no
longer creates UNIX sockets.