testserver: change `firstTenantID` to 3 due to UA migration
Following the migration to UA, we are not allowing tenantID 2
to be used as an application tenant anymore. Currently testserver
creates tenantid 2, but it will fail to do so in the future
following the rule to disable tenant id 2 in
https://github.com/cockroachdb/cockroach/pull/148561.

This change will allow us to start assigning tenantIDs from 3 and
pass the failing tests in #148561.